backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
ariadne==0.26.2
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
ecdsa==0.19.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.1
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
